Split Connector into TcpConnector and DnsConnector
